--- src/components/Admin/SchoolManage/index.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const SchoolManage = () => {
-  return <div>SchoolManage</div>;
-};
-
-export default SchoolManage;
